requirements.txt
einops==0.6.1
lifelines==0.27.7
matplotlib==3.7.2
nibabel==5.1.0
numpy==1.22.3
numpy==1.24.4
opencv_python==4.8.0.76
openslide_python==1.3.0
pandas==2.0.3
Pillow==10.0.0
Pillow==10.0.1
pymrmr==0.1.11
pyradiomics==3.1.0
PyYAML==6.0.1
PyYAML==6.0.1
scikit_image==0.18.3
scikit_learn==1.0.2
seaborn==0.12.2
SimpleITK==2.2.1
SimpleITK==2.3.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
